requirements.txt
Flask
gunicorn
python-barcode
Pillow
